view log/ clear log

+++ logger.py
@@ -1,3 +1,4 @@
+import os
 from datetime import datetime
 
 #Create the logger class to helo the user track their goals/ needs
@@ -5,9 +6,10 @@ class Logger:
     def __init__(self, filename="mindfulness_log.txt"):
         self.filename = filename
         # Overwrite file at program start
-        with open(self.filename, "w") as f:
-            f.write("Mindfulness Log\n")
-            f.write("====================\n\n")
+        if not os.path.exists(self.filename):
+            with open(self.filename, "w") as f:
+                f.write("Mindfulness Log\n")
+                f.write("====================\n\n")
 
     def log(self, exercise_message):
         """Append a completed exercise with timestamp."""
